+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Keyboard Macro Bata extension will be documented in this file.
 
+### [Unreleased]
+- New
+  - Added Sublime Text Keymap support. [#18](https://github.com/tshino/vscode-kb-macro/issues/18)
+
 ### [0.5.0] - 2021-12-14
 - New
   - Added Web extension support. [#17](https://github.com/tshino/vscode-kb-macro/pull/17)