+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to the Keyboard Macro Bata extension will be documented in t
 ### [Unreleased]
 - New
   - Added Sublime Text Keymap support. [#18](https://github.com/tshino/vscode-kb-macro/issues/18)
+- Fix
+  - Keymap wrapper for Awesome Emacs Keymap may have not been working correctly on macOS due to "mac" keys in user `keybindings.json`.
 
 ### [0.5.0] - 2021-12-14
 - New