+++ CHANGELOG.md
@@ -15,8 +15,8 @@ All notable changes to the Keyboard Macro Bata extension will be documented in t
   - Added Web extension support. [#17](https://github.com/tshino/vscode-kb-macro/pull/17)
   - (Internal) Introduced Conditional Await syntax on keymap wrapper's config file.
 - Update
-  - Updated keymap wrappers for Vz Keymap; Reduced unnecessary delays in playback.
-  - Updated keymap wrappers for Awesome Emacs Keymap; Reduced unnecessary delays in playback.
+  - Updated keymap wrapper for Vz Keymap; Reduced unnecessary delays in playback.
+  - Updated keymap wrapper for Awesome Emacs Keymap; Reduced unnecessary delays in playback.
 
 ### [0.4.0] - 2021-12-10
 - New